New Filemanager system installed. This is used only for input files that are declared in the rundeck. 	- FILEMANAGER module in UTILDBL.f 	- In rundeck, input files now have names, not numbers 	- setup links to the name, not to fort.x 	- in the GCM, a unit number is assigned to a name on a first 	  come, first serve basis using the 'call getunit' before the file           is read. 	- 'getunit' opens the file 	- multiple units can be requested at a time (call getunits) (this 	  should really be the same as getunit, but it seems a little tricky, 	  any ideas?) 	- FILEMANAGER keeps track of how many units have been used and so 	  always ensures a unique number (starting at 50 to avoid clashes 	  with output files).
Consequences of this are that you never have to worry about clashes,
and that any number of input files can be added with no additional
changes to the code. 'setup' still knows that some files
(i.e. AIC,OIC,GIC) are I.C. and can be left out of the link file, and
otherwise works identically as before. Even the radiation files are
incorporated into the system. Let me know if you like it.

Be sure to 'mkrun E001 E001xyz' to get the necessary changes to the rundecks.