Constants have now been made double precision (at least in const.f), and most redefinitions of constants have been eliminated. Current 'best estimates' of various constants have replaced the previous definitions (esp. for rvap,rgas,grav and kapa).
Some recoding in PGF to take advantage of previously defined
reciprocals. Please continue to eliminate duplicate definitions as
you notice them.

Note that this significantly changes the rsf/prt files. During the
first hour, there is a large change as the fields adjust to new values
of fundamental constants.  For instance the planetary albedo decreases
from 32.85 to 29.89. DO NOT be alarmed, it is just a short term
thing. After one month the changes do not seem as significant
(planetary albedo now 31.80 compared to 31.87 previously. The rsf/prt
files have been updated in /u/cmrun/modelE.