Implementation of local i/o routines.
I have moved all i/o related processes to local subroutines, one for
each sub-module. They are generally found immediately after the
decalrations of the prognostic variables (i.e in RADNCB.f or
CLD01_COM_E001.f etc). The idea is that local changes to the rsf file
should not require changess to the rest of the model.

Note that options during the rsf file read (that are used for the
various ISTARTs) are now fully defined in the i/o routine
itself. Currently there are three: irestart (reads everything in),
irsfic (doesn't use the diagnostics, the C, JC, RC arrays, or some of
the radiation arrays), irerun (uses some but not all information).

I have added S0 to the restart file (in the radiation section - along
with RSDIST,COSD,SIND), to correct a restart problem wit ADAILY(1).

Also, I have replaced ODATA with TOCEAN(3), RSI and MSI. TOCEAN belongs
to the OCEAN module, while RSI and MSI belong to SEAICE_COM.

As an experiment, I am continuing to reshape the arrays used
exclusively in column physics calculations so that the L dimension
comes first. I think that this has improved things in radiation a
little but it is very hard to tell! Other candidates would be the
ground hydrology arrays (wbare etc.) and the second order moments.

PRT files are the same, rsf files are different (they have been
updated). Please 'make cmp' to update the CMPE001 program.