Modifications to Makefile:
Switched from SGI "make" to GNU "make". Now you have to use "gmake"
instead of "make".
Included multi-system support: SGI, IBM and Linux.
"setup" at the moment works only on SGI, though it should be easy to modify
the "setup" script to make it work on IBM workstations.
Linux support means that if we get commercial version of Vast f90
compiler then we should be able to compile gcm on Linux.

Makefile now is more user-friendly. Type "gmake" without arguments to see
a short help message. Dependencies are now generated autamatically, so
you shouldn't do "gmake depend RUN=..". You can also skip other targets.
One piece of weirdness is that if you generate a new rundeck with
"gmake mkrun RUN=...", then the message "Regenerating Rundeck..." will
be repeated two times. This is a consequence of automatic generation
of dependencies and I don't know how to get rid of it and preserve
current functionality (I could of course just remove all messages :-) ).

To make rundecks machine-independent I introduced a meta-character "~".
This character if present in a rundeck inside a name of an object file
will be replaced by the Makefile with the system type on which it is
compiled (SGI, IBM or Linux). This is basically necessary only for the
files RANDSGI and RANDIBM, so that they can be chosen automatically
on different architectures.

File "mkrun" was removed since it is not used any more.

Script sfmakedepend was moved to subdirectory ./SCRIPTS