Modularization of PRECIP/GROUND
I have now mostly finished modularising PRECIP and GROUND. These
routines have now been replaced by calls to individual drivers for the
sea/lake ice, the landice, the lakes, the ocean and earth (only
diagnostic). The order in which they are called is important. The sea
ice driver, and the landice driver must be called first since they
calculate fluxes that are then passed to the lakes or oceans.

As far as possible, the sea ice code and the ocean code are now
completely seperate, as is the code for lakes and oceans. The drivers
can be found in two new files LANDICE_DRV and SEAICE_DRV. I've put the
lake and ocean ones in LAKES and OCNE001 for the time being.

I have added a module FLUXES that holds most of the things that are
fluxed betweeen the atmosphere and other components (possibly precip
belongs here too). Note that we currently save delta T/U/V/Q values
for the SURFCE fluxes. These should probably be changed to the flux
quantity itself.

Note that the rundecks have changed to accomodate the new files.

The recent series of bugs from the seaice and Qflux models have been fixed.

rsf/prt files will be updated.