Recasting of AJ zonal diagnostics to work over multiple types.
BJ and CJ arrays have been superceded. AJ is now (JM,KAJ,NTYPE). The
number of types, and their definition is set in E001M12_COM. Currently
I have defined 6: ocean/ocean ice/lake/lake ice/earth/land ice.
Much of DIAGJ was changed to incorporate this.

I have added one diagnostic 'AJ(J_TYPE)' which saves the surface
fraction of each type. This is a little awkward to add to the budget
pages (because it can't be weighted by it's own surface fraction), but
the fudging is minor. AJ(J_RSI) remains, as before, the amount of ice
over lakes and oceans. Possibly this is now obsolete.

The output for the budget pages is now more voluminous. 9 categories
are used. Each type is done seperately, plus the global numbers, and
also all ocean and all lake composites. This can be easily changed if
required.

Note that one diagnostic (MC CLOUD DEPTH = J_CTOPP) is clearly
incorrectly specified (the summation is of a ratio!) and changes with
the definition of the number of types. (Anyone want to fix this?).

I have added the array 'FTYPE(NTYPE,IM,JM)' which carries the fraction
of each surface type used by AJ. This could be (should be?) added to
the ACC file for diagnostic purposes. This array MUST be updated
whenever there is a change of a surface fraction (currently this is
only when RSI changes, or (less often) when FLICE changes).  Note that
this array can be used for giving surface fractions ONLY within a
1,NTYPE loop. It cannot be used as a replacement for POICE etc. (this
is becasue of the possibility of overlap between types in the
diagnostics, but not in the prognostic code.)

In order to ensure that the separate types exactly add up to 100%, I
added some code to INPUT that forces it to be true. (previously, it
was only true to single precision).

These changes have led to a great deal of array-ification of the AJ
accumulations which have simplified a lot of the code in the driver
routines and DIAGA. J_EPRCP accumulation moves to CLD01_DRV, and
PRECIP_E is merged with GROUND_E (which in turn should be merged into
EARTH). Possibly some further tidying is possible.

Further work is required to generalise DIAGJ as much as I would like
it to be.

The rsf/prt files have been changed.