Expansion and increased flexibility for conservation diagnostics
I have rewritten the conservation diagnostics so that they can be
added to in an automatic manner (with one call to the new routine
'get_ncon'). As an example I have added a lake mass and lake energy
diagnostics, and I expect soon to add the ground hydrology, oceans and
ice mass and energy budgets..

DIAG9A is now much more straightforward, with all of the mechanics
handled by conserv_DIAG and DIAG9P (which now do not depend on the
conserved quantities at all). Angular momentum and kinetic energy are
special cases because they are also set within the dynamics.

Each conserved quantity is now calculated by a specific routine that
returns only the zonal averaged amount (see for example conserv_AM in
DE001M12.f).

rsf files are now slightly larger, and the prt files differ due to the
extra printout. The controls have been updated.