Makefile changed to put CMPE001 into the current directory (aux)
CMPE001.f changed to distinguish between (I,J,L) and (L,I,J) arrays
and to account for unused entries (poles). (Both goals were only
partially achieved)

Note: If changes are made that change the results, please do at least
the following test:  run the model for 2 hours (set HOURE=2,NDISK=1, in
the rundeck), make a copy of the final restart file, rerun the second
hour and compare the new restart file to the copy of the old one.
Except for the timing numbers, the 2 files should be identical.
(to run the 2nd hour, add in file I  the line
      ISTART=99,YEARE=1950,MONTHE=1,DATEE=1,HOURE=2, )

Unfortunately, our current model does not pass that test. I will wait
with adding a correction to the radiation until that bug has been
found and corrected.