Names, units, titles for accumulation arrays added to DAGCOM
These quantities are set in new file DEFACC.f (added to rundeck).
Many names, units, titles, etc. need to be completed...

There are overlaps with existing NAME_IJ, etc. because
DEFACC.f was not originally contructed for modelE.

New file NCACC.f added to model directory, but not
added to rundeck.  This code writes acc-files in netcdf format.
Currently it is not invoked by the main model, but rather
by auxiliary program acc2nc (added to aux directory; it
converts existing acc-file format to netcdf format).
To have the main model write acc-files in netcdf format,
add call to 'write_nc_acc' in main program, add NCACC to
rundeck, and link to netcdf library when executable is
created.  Note that adding NCACC to the rundeck produces
the undesirable side effect that gcmlib.a refers to
the netcdf library which must then be linked to by all
auxiliary programs invoking gcmlib.a, even if these programs
do not explicitly invoke the netcdf library.  Suggestions
about how to avoid this 'kitchen-sink' behavior of the
linker are desperately needed!

DEFACC not invoked by GCM yet, but rather by NCACC.

New programs in 'prtdag' directory read netcdf acc-file format
and write selected diagnostics to netcdf format.  Because the
acc-files they read are self-documenting, these programs are
logically independent of the model.
These programs are limited in scope, they do not write all
existing diagnostics yet.
Lots of units, titles, etc. need to be filled in...

A script to sum acc-files will be added soon.