New predicitive lakes component:
This is a substantial change to the lakes in the model. The files
LAKES.f and LAKES_COM.f are essentially complete replacements for the
previous versions. If you have modified (and not yet committed) either
file please do not attempt to commit until you have talked to me.

Summary of changes:

- New two-layer thermodynamic lake model. Original prognostic
variables (MWL,GML,TLAKE) are retained, and a new variable MLDLK (the
depth of the first (mixed) lake layer is added.

- Lakes now completely divorced from ocean code. Predicted lakes will
be used even for climatological SST runs.

- Minimum ice thickness has been reduced to 20cm (from 50cm) - this is
used for ice over the ocean as well as for the lakes.

- Additional diagnostics to track ice-on/off dates for lakes, using an
extened version of TSFREZ. Two new AIJ indices are used to store these
diagnostics for accumulation. This could also be extended to the
sea ice fraction.

- T50 obsolete and has been removed.

- Correction of a scaling error in thr river_diag routine. (Runoff was
10 times too large in printout).

- Changes to SURFCE/SEAICE necessary to prevent lakes drying out.

- Initial start to code that would allow FLAKE to vary in
future. Thus, only FLAKE0 (the original value) is read in from the
topography file. FLAKE is now a LAKES variable, and could be adjusted
every day if required. However, the conservation of
energy/water/tracer in the land component needs to be thought about
before this can be implemented. Diagnostics that assume fixed
FEARTH/FLAKE also will have to be adjusted.

Effects of change:

Test runs have gone three years with no problems. Anomalously warm
lakes at high altitude have disappeared. Ice-on and ice-off dates for
high latitude lakes seem consistent with observations. Further analysis is
warranted.