"gmake setup RUN=..." now moves the executable from BIN_DIR instead of copying it. This prevents double copies of the same file.
A new target "alias" is introduced, which would link executables
"qc" and "CMPE001" for specified rundeck to current directory, i.e.
one does it with
   gmake alias RUN=<my_run>
to remove them do:
   gmake unalias
I think that is really a bad idea, because those files may be
rundeck-dependent. But since some find it useful we may keep it
in the development version of the model.