Sea-ice changes:
1) Solar radiation is allowed to penetrate ice (and through to the
ocean/lake if necessary), This is new physics and has the effect of
reducing ice surface temperatures, and surface melt. Snow cover and
snow depth increase in the melting hemisphere. (Note that the routine
SEA_ICE was completely rewritten). This may prove to be a reasonably
big effect.

2) The treatment of the ice for the fixed SST/seaice runs has been
changed.  Now, all fluxes due to PRECIP or SURFCE are calculated and
applied, and any consequent change in MSI2 is matched by an implicit
flux from the ocean (which brings in ice at the ocean freezing
point). This does not appear to have made a huge difference.

3) Two minor corrections have been made: i) The open ocean fraction
(controlled by FLEAD), occasionally produced sea ice fractions > 1 if
the ice was thicker than 5m. ii) There was a small inconsistency in
the SEAICE code regarding DEW being applied to the ice.  These have
been fixed.

A couple of diagnostics have been redefined (ie. J_DIFS and
J_EDIFS). These are now the implicit mass and heat fluxes at the base
of the ice, not at the first-secopnd layer interface.

Note also that the ocean temperature used to calculate the flux
through the sea ice (in GROUND_SI) is now the freezing temperature of
the water, and not the actual temperature. Again a minor point.