Since some had problems with the number of open units supported by FILEMANAGER, I went ahead and modified it so that now it recycles closed units. New subroutine was introduced called "closeunit". It should be called instead of "close". It closes the file and also releases the number of the unit so that it can be used again. To make the names more conformal I renamed "getunit" to "openunit". I left the interface of "openunit" unchanged except that two last arguments are made optional. If omitted they are considered to be .false. (i.e. "openunit" will open "NEW" "FORMATTED" file). This default behavior is convenient to open text files for debugging output. So typical usage of "openunit"/"closeunit" would be:
     call openunit( "some_file", iunit, qbin, qold )
     ... read/write something from/to unit <iunit>
     call closeunit( iunit )

Also "getunits" was renamed to "openunits". "closeunits" was modified
to take as an argument an array of units to be closed. Hence "openunits"
and "closeunits" are array versions of "openunit" and "closeunit".
A function "nameunit" was introduced which given the unit number returns
the name of the file to which it corresponds (actually last 16 characters
of it).

I went over the model renaming "getunit" and inserting "closeunit"
where seemed to be proper.
I also did the same to dec31.E001.f otspec.E001.f vertflux.E001.f and
also fixed declarations for iu_* (but somebody more familiar with this
code please check if I resolved conflicts with global names in
MODULE OCEAN properly).

Restart files not changed.