Re-configuring and correcting of the ocean related code.
I have now split OCEAN.f to allow the normal Qflux model and the deep
diffusivce Qflux model to work as plug and play options. For the
standard Qflux use: OCEAN and OCNML, for deep diffusion use OCEAN and ODEEP.

For the coupled ocean I have added in support for the diagnostics
along the same lines as for the atmosphere. No print out as yet
though.

Minor changes to CONST and the DIAG files.