I know I shouldn't have, but I have edited the common blocks in BR00B.COM to separate the variables that are set in BLOCK DATA, from those that are read in, and those that are not initialised at all. Some unnecessary initialisations in RADIATION.f were also commented out.
This has the effect of reducing the size of the object file for RADIATION.f
by a factor of 7  (from almost 8MB to less than 1MB). The executable
for the model also reduces from 8632720 to only 1918696. This should
make it easier to compile on smaller systems.

Running for a month shows no differences. However, we should keep an eye
on this.