Makefile functionality slightly changed - PLEASE READ (at least next ten lines ...)
Added functionality to keep global log of all runs being done.
It works as follows:

When you create a rundeck with
    gmake rundeck RUN=my_run
the Makefile checks if a rundeck with such name is already present in
global repository.  If not then the rundeck is created and copied to
global repository.  If yes then Makefile complains and exits without
creating the rundeck. If you want to create it anyway and overwrite
the one which is in repository you have to do:
    gmake rundeck RUN=my_run OVERWRITE=YES

When you do setup with
    gmake setup RUN=my_run
the Makefile overwrites the rundeck in the repository with its current
contents and appends to it the following info:
   - current date and time
   - current compiler version
   - the name of your CVS branch

If you recompile executable later with (presumably to continue the run)
    gmake exe RUN=my_run
the Mkefile will append to the rundeck in repository the following info:
   - current date and time
   - current compiler version
   - the name of your CVS branch
   - "diff" between rundeck in repository and your current rundeck

This functionality should provide enough info to reproduce the run
at later time if necessary. There were objections against centralized
naming of rundecks, so such machinery is not implemented. It is
responsibility of each user to choose reasonable names for rundecks
and keep track of them. One should not use of course the same name
for two different runs ("make setup" will overwrite information on
previous run with the same name).
Please let me know if you think that some other vital information
should be included into global log.

The global rundeck repository is now set to: /u/cmrun/modelE/decks