Formally separated the calculations of the heat diffusivity kh and the moisture diffusivity kq (but at present, still use kq=kh). AS a result, the common block PBLOUT contains ...km,kh,kq,..., instead of ...km,kh,... A better moisture model will be put in in which kq and kh are not the same.
SURFACE.f and GHY_DRY use this common block, how
to use kq when kq and kh are different
in these subroutines may pose a question.

In PBL.f and ATURB.f the diffusivity coefficient, sq, for
the transport of the turbulent kinetic energy,
is adjusted. A more complete model will be replacing
the current simple model, thus improving the nonlocal diffusion.

Ye
CVS