All 'open' statements consolidated so that it only appears in FILEMANAGER.  (except one used to control the standard output in DIAG.f). This was done so that compiler dependent options can be kept to an absolute minimum.
In RADIATION.f all superfluous OPEN/CLOSE stamements commneted out

In UTILDBL.f added preprocessor option 'CONVERT_BIGENDIAN' to add an
extra statement to the OPEN commands if a compiler requires this to be
able to read UNIX style binary data files on a PC platform.

In Rules.make I have attempted to make the Rules compatible with Linux
and SGI, plus I have added 'CPPFLAGS' in order to pass variables to
the C-preprocesor based on the compile environment. This has meant a
slight (!) increase in the complexity of the compilation commnd line,
and so I have replaced it in the output with a
'Compiling PHYS.f...Done' statement instead.

In PBL.f minor double precision changes