Diagnostic reform: These are a whole set of changes to the diagnostics that will helpfully improve peoples understanding of what is being output. This is a first cut, and so I expect some modifications to occur as people look at their favorite diagnostics.
The changes are mainly to the scaling and units for each diag. (Mainly
the JK/JL diagnostics). We tried to keep to the following principles:

i) No resolution dependent units. (i.e. the output should look roughly
the same regardless of horizontal and vertical resolution).

ii) Units should be comparable to observations, or commonly used
conventions.

iii) Units of similar quatities should be easily comparable.

Of course, some of these goals are contradictory, but hopefully the
compromises we have come to will be ok for most people. Thus all the
/DSIG units have been replaced with /mb, JOULES/UNIT SIGMA replaced
with m^2/s^2 etc.

There is now an additional argument to JLMAP and JKMAP which is a
power scaling. This scale no longer appears in the binary file (which
will be in the raw unit), but it does affect the printout.

Note that the DP that is used in the denominator in JKMAP is the
accumulated pressure, not the accumulated SIGMA. This implies that the
horizontal averages are now properly mass weighted, and therfore
slightly different to those before.

I have also tried to enforce some consistency with unit naming conventions.

Please let me know if there are any problems.