Additional diagnostics:
i) I have included an option to calculate some cloud diagnostics
according to the ISCCP algortithms. These diagnostics are calculated
in CONDSE and can be set using the option 'isccp_diags=1' in the
rundeck parameter listing. The default is not to calculate them.  If
they are selected, then of the three extra six diagnostics are
included in the printout (LOW, MIDDLE and HIGH CLOUD COVER) at the
same time as the original diagnostics of these quantities. Note the
new diagnostics have (ISCCP) in the title to distinguish them. There
are also CLOUD TOP PRESSURE (ISCCP), CLOUD OPTICAL THICKNESS (ISCCP)
which appear in the pdE output.

Note that the ISCCP versions have significantly less low cloud cover
than the standard diagnostics (for one month, 14% compared to 43%),
mainly due to the sheilding effect of higher clouds.

ii) I have also added an IJ diagnostic for the tropopause pressure