I have added a version of Gary's RMS program to the aux directory. This takes lat-lon diagnostic output, and calculates the RMS differences to selected observations. I have been taking the last 5 years of a 6 year climatological run.
The results are stored in /u/cmrun/modelE/RMS (currently hard-coded),
one file for each model run.

Please make use of this utility to judge whether changes you have made
actually improve things!

Note that a link to the TOPO (topography file) must exist in the
directory where this is run.

Currently there are three runs stored: E001gas, E002gas and E004gas,
which represent the last few development stages. E002gas corresponds
to modelE-1-3, and E004gas is a test version I am currently working
on.

It would be nice to have a script that takes a list of model runs and
makes a concatenated list for easy comparisons.