Fixes for some outrageous lapses in programming (mine unfortunately).
i) in icelake, the mass flux 'm' was never set (but was used!)
ii) SSIDEC was only called when there was nothing to do
iii) SOLAR(2) only accumulated half of the surface time steps.

Of these errors, only iii) affects model versions 1-3-1 and
earlier. It has the effect of making more solar radiation be absorbed
in the first layer sea ice (instead of penetrating down). Since this
is quite close to the original coding anyway, there is no need to
alter older code.

After one month with the fixes, differences are relatively minor (at
least for the fixed SST case).