Proper initialisation of some ice dynamics variables.
The sea ice concentration gradients used in ADVSI are not referenced
by the rest of the model, so I've put a gradient check at the start of
ADVSI so that we don't get bizarre advection errors.

I think that this will cure the coupled model problem, but please let
me know if that isn't the case.