Some more fixed problems with sea ice and ocean code:
i) In OCNDYN set GTEMP array for sea ice as well since ice can be
advected through straits and can change seaice amount. This previously
caused a small error every radiation time step since GTEMP was not up
to date at a small number of points. This also caused a
reproducibility error if you restarted at a radiation calculation.

ii) In SSIDEC the salinity was decreased incorrectly in some layers
which gave rise to negative seaice salinities.

iii) In UNDERICE the basal mass flux occasionally exceeded the mass of
the base layer, which led to possible errors in the heat and salt
fluxes. Mass fluxes are now capped at 90% of the mass of bottom
layer. This seems to be important about once a time step.