Removed SURFACE_turb.f - instead modified SURFACE so that the part of the code which was removed in SURFACE_turb is called as a subroutine. This subroutine was put into DRYCNV.f and dummy with the same name was put into ATURB.f. E001*.R modified to previous state.
Also some extra code was added to PBL.f to modify boundary condition for
the q-equation at the bottom to:
!@+   kq * dq/dz = min ( cq * usurf * (q - qg) ,
!@+          fr_sat * cq * usurf * (q - qg) + ( 1 - fr_sat ) * flux_max )
where fr_sat is fraction of saturated soil, and
      flux_max is maximal evaporation flux from unsaturated soil
This was necessary to do proper flux limiting for evaporation from the
soil. Flux limiting code is not yet properly tested and is not included.
For now fr_sat is set to 1., which corresponds to saturated soil and
disables flux limiting. For evaporation from water/snow/ice fr_sat should
always be 1.
PBL_DRV.f converted to a module to make it easier to pass input/output
parameters.