Addition of two new database parameters and the ability to control the concetrations of CH4 and H2Ostrat for the radiation from the rundeck.
Both parameters are simple scalings over the default value (similar to
the CO2 parameter, when that is positive). Default's are 1.

H2Ostrat only scales water concetrations above the tropopause
(LTROPO(I,J)+1,LM), and thus the tropopause level needs to be passed to the
radiation using a new parameter (LS1rad) set in RADIA.

Look at the printout for the current values of CH4, CO2 and scale
accordingly (note that there is a latitudinal gradient in CH4). The
base concentration of H2Ostrat for the 12 layer initial concentrations
is 2.5 ppmv (possibly a little low).