More tracer code enhancements:
1) tracers now work with DRYCNV or ATURB (although non-local terms -
which are currently not yet standard still require work).

2) Conservation diagnostics now have one extra option for passing a
'delta' tracer field, as well as simply looking at the difference
between now and before. This has been coded so that separate surface
sources can be clearly enumerated as well as the separation between
moist convective and large scale condensation changes. The sum total
change between routines is then a check that the individual DELTA's
are being done correctly. An example output:

 ---------------------------------------------------------
                                         GLOBAL     NH
 ---------------------------------------------------------
0INSTANT Water (10^-3 kg/m^2)          21231.23 18518.95
 CHANGE OF Water BY DYNAMICS               0.00 -1874.32
 CHANGE OF Water BY CONDENSATN        -22507.06-24598.01
 CHANGE OF Water BY LAND SURFC         26074.12 32772.56
 CHANGE OF Water BY SURFACE                0.00     0.00
 CHANGE OF Water BY FILTER                 0.00     0.00
 CHANGE OF Water BY OCEAN                  0.00     0.00
     DELTA Water BY MOIST CONV         -6879.47 -8350.18
     DELTA Water BY LS COND           -15627.59-16247.84
 SUM OF CHANGES (10^-9 kg/s/m^2)        3567.07  6300.23

Note that the two DELTA's add up to the total CHANGE by CONDENSATION.
Some superfluous diagnostics for the test cases were removed.

3) Water (and hence sea ice) tracers work properly if the coupled
ocean is turned on whether or not the ocean tracers are being
advected.