runpmE modified so that it restarts the model with a smaller time step only when it stopped with a specific retcode (retcode = 11). This will prevent the model from restarting when it shouldn't.
sswE algorithm changed from file reading to signal handling. "fort.3"
is not created any more. sswE just stopes the model with "kill -15".
Appropriate signal handling was added to the model so that it finishes
gracefully at signal 15. This approach is more UNIX-friendly and
it doesn't slow down the execution (as opening/reading/closing file
does).
This change should be transparent to everybody: you still use
"sswR RINID" to stop the model. If you use "kill -9" on the model it
will stop immediately as before.