Retcodes = 11 are put into places in QUS_DRV.f where the program stops due to big time step.
Creation of fort.3 is removed. Instead a logical variable OFFSSW=.false.
is introduced, which is set to .true. when a signal=15 is received.
The value of this variable is used to stop the model.

PS new "sswE" is put into /u/exec. It is not compatible with older runs.
Please use sswE.old for older runs.