And they said it couldn't be done.....
Yes, ladies and gentlemen, roll up to see, the one, the only, set of
conservation diagnostics for AM and KE that actually add up!

Gasp at the comparison of the internal dynamics diagnostics with the
integrated change!

Thrill when they are exactly the same (for AM) or close (KE)!

Swoon at the individual accounting of SDRAG, the UV Filter and the
Gravity wave drag!

Yes, all this can be yours! Update today!