Diagnostic changes:
- I have redefined many of the heat and mass flux diagnostics in the
budgets pages. Essentially, I have removed fluxes that are internal to
the sub-modules (like the flux at Z1), and replaced them with the net
fluxes between modules (i.e. the runoff, river flow, implicit fluxes
etc.) that are more important for the overall mass and energy budgets.

- Note that for balance with an ocean model, the OCEAN global value of
"NET HEAT Z0+RVR" must be close to zero. This is roughly 1.43 times
the GLOBAL value due to the surface ocean fraction, but can vary by a
watt or two due to various land-based dis-equilibria.

- Removal of indexing in DIAGJ. Now j diags are printed in the order
they are defined in DEFACC.f.

- Removal of irrelevant diagnostics from the REGIONS budget page

- Correct calculation of the moist convective cloud depth (which was
incorrect because of the summation over type of a ratio). This is now
treated like an albedo.

- New jl diagnostic from CLOUDS, "EFFECTIVE RELATIVE HUMIDITY", saves
the actual humidity that is compared to U00.

- some minor rewriting of some SEAICE routines

- Removed misleading comments concerning diagnostics in DIAG.f

- some correction and addition of regional diagnostics

- general tidying up of diag code.

**********************  NOTE FOR OCEAN MODELS: ***********************

The budget diagnostics in the OCEAN code have changed, please ensure
that any analagous code in other ocean models takes that into account.

**********************************************************************