A number of fixes and simplifications for the ocean and ice dynamic models:
ICEDYN:
 - corrected mistake in application of north pole sea ice fluxes
 - removed some unecessary arrays
 - correct scaling error for ice-ocean stress

OCNDYN:
 - change to OPFIL to work on varying layers
 - OPFIL now used to smooth sea surface height prior to being used in ice
   dynamics

OCNGM:
 - fixed spelling mistake

As far as I can tell, this removes an instability that occured in the
Arctic ice and ocean velocities. This appeared to be a problem when an
initial noise pattern in the ocean was amplified by the movement of
the ice above it. I think that this has now been eliminated, but you
never know...