Upgrade of ocean diagnostics for Gary's model. Now includes print out of mass/heat/salt transports and new lat-lon diagnostics in binary files. Latitude-depth and longitiude-depth diags (streamfunction, temp/salt sections and means etc.) are now output to binary files (MONYEAR.ojlRUNID and MONYEAR.oilRUNID) using POUT_JL and POUT_IL. Full transports including breakdown by basin and overturning vs. horiz. gyre vs. GM flux are available in MONYEAR.otjRUNID.
This will work with all current runs if you 'gmake exe RUN=...', and
use existing acc files.

Unfortunately, some necessary OIJ diags are not currently saved and I
have to resort to using the atmospheric versions. This will be
corrected in a later update, but that will cause the diagnostic arrays
to change, and hence will not be backwards compatible.

Tracer diagnostics still not included....