*************** IMPORTANT CHANGES TO CLOUD SCHEME ********************
One of the last changes that was made to the clouds was to alter the
calculation of the critial relative humidity threshold for clouds to
form in the plantetary boundary layer. For the 12 layer model this
made a 4 W/m^ difference to the surface heat balance, and lead to
difficulties with the overall radiative balance. However, it did
appear to imporve the cloud distribution in the 18 layer model.

This change now re-does this calculation. Instead of basing the
crit. RH on each individual layer thickness, it is now based on the
integrated thickness from the ground. This clearly favours cloud
formation at the top of the PBL, and is more likely to remain
consistent as layers are added to the lower part of the model.

Preliminary tests show that the previously anomalous effects on the 12
layer model have disappeared, while the beneficial aspects to the
cloud distribution in the 18 layer model are maintained.

NOTE: This does have a significant impact on the total radiation
balance, and so models will require retuning. The 12 layer model seems
ok with the U00water=0.49 and U00ice=0.5 as in the default rundeck,
preliminary estimates for the 18 layer model indicate a U00 near 0.56
is likely to work.

**********************************************************************