Important error found in ice-ocean fluxes for fixed-SST and Qflux model =======================================================================
It appears we were double counting the cooling of the ocean under the
ice (and thus the oceans would cool too much). This is not an apparent
error for the fixed SST case since the mistake will be compensated for
by the daily restoring, however in the Q-flux model it would lead to
erroneous heat losses and gradual cooling of the climate.

Note that the Q-fluxes will need to be re-calculated once this fix is
applied.

This error may have existed since the end of last year. Sorry about that.