1) The melting of small amounts of sea ice (rsi<1d-4) in daily_ocean and daily_ice, has been replaced with a call to MELT_SI prior to the surface flux calculations. The fluxes from this melting are now stored in the FLOWO,EFLOWO arrays with one new array SFLOWO (salinity in the sea ice melt). This has been done to further modularise the ocean code.
IMPORTANT: ALL OCEAN CODES NEED TO TAKE THE NEW ARRAY SFLOWO INTO ACCOUNT.

2) The function TOFREZ(I,J) has been replaced with TFREZ(SSS) which
calculates the freezing point of sea water as a function of the
surface salinity. Again this makes life easier for the ocean models
which now do not need to worry about TOFREZ.

3) Some technical diagnostic fixes in RIVERF (no effect)

4) Added diagnostics to GLMELT to improve mass balance calculations
for coupled ocean code. Minor tidying up.

5) Found a minor error in fluxes of river runoff to ocean in Gary's
ocean code. Not sure how big this will be, but it had the effect of
not putting enough mass into the ocean from rivers.