Sub-daily diagnostics:
We finally have the capacity to produce timeseries of selected
instantaneous quantities at sub-monthly frequencies. This can
therefore produce hourly, 3-hourly, daily etc. diagnostics as a
function of the NSUBDD database parameter (set in rundeck). NSUBDD=1
gives hourly output, =12 twice a day, etc.

The diagnostics are selected using the string SUBDD in the
rundeck. This string is a space seperated list of the required
diags. Currently the options are SLP, SAT, QS, PREC, Z500 (500 mb
geopot height). Other options can be added by adding cases to the
routine get_SUBDD in DIAG.f. Multiple options (up to 10) are easily
configured in the rundeck, although the total string length of SUBDD
is limited to 32 characters.

For each selected diagnostic, a timeseries file NAMEMONYEAR is opened
each month (if SAT is selected, the first file will be SATJAN1950).
The file will contain one month's worth of the field and the time
information 'it,DATA,it', saved every NSUBDD hours. If a model is
being restarted it will find the appropriate files and append new
records as necessary. The file format may be improved upon at a later date.

This coding supersedes the 'NSLP, get_SLP' functionality, which can
simply be reproduced using NSUBDD=12, SUBDD="SLP". All code has been
placed in the subdaily module in file DIAG.f.

All rundecks have been updated to reflect the change.