Adjusted River runoff routine to allow river flow from land to ocean in same box.
New rules:

i) All boxes that are 100% land must have an explicit river direction

ii) Boxes which are a mix of land and ocean have as a default that the
land runoff goes into the ocean in the same box. (But this can be
overridden by including an explicit direction).

iii) Named river mouths in the river direction file must be in boxes
with some ocean (and the runoff from any land fraction in the same box
goes into the ocean there. This cannot be overridden).

If these conditions are not met, warning messages are printed.