BIG CHANGES FOR ICE DYNAMICS:
The ice dynamics code has been completely uncoupled from Gary's ocean
code. This will make it much easier for other ocean codes (or even the
q-flux model) to include ice dynamics.

However, this comes with a small price. Since icedyn is now
independent (and can have a different grid from the rest of the
model), all entry points, diagnostics etc. have to be specifically
placed. Thus we now have new routines which MUST be included in ANY
code. For the fixed SST/q-flux models, they are dummy routines, and
for any ocean that does not want ice dynamics, the same should be
done.

Currently, the ice dynamic grid is on the atmospheric grid. This can
be altered in two stages. Firstly, the ice rheology/velocities can be
calculated on a higher resolution grid (i.e. closer to the ocean
resolution) while the advection (in ADVSI) stays on the atmospheric
grid. Secondly, ADVSI could be reworked to perform advection on the
higher resolution grid as well. Note that any new grid must be a
spherical lat/lon projection.

New routines:
     init_icedyn, diag_icedyn, io_icedyn, io_icdiag reset_icdiag

Summary of changes:

 - New split between ICEDYN.f (local routines to ice dynamics only), and
   ICEDYN_DRV.f (which contains the entry points DYNSI,ADVSI,and all
   new routines).

 - OIJ diag arrays for ice dynamic quantities are now ICIJ
   arrays. INstead of being output by diag_OCEAN, they are now output
   with diag_icedyn.

 - Ice dynamic grid is set by IMIC,JMIC. Can be changed as long as the
   highlighted interpolations in ICEDYN_DRV are altered also. All diags,
   and local code are indpendent of the actual grid.

 - IORSF now contains specific ice dynamic entries.

 - MODELE to simplifiy things a bit, I have created a new routine
   'print_diags' which is called from various points in MAIN.

 - OCNML/ODEEP new dummy routines.

 - OCNDYN/OSTRAITS removal of anything to do with ice
   dynamics (including commenting out STADVI, which never worked
   properly anyway)

 - FLUXES new flux used in ice dynamics: OGEOZA which is the ocean
   geopotential on the atmospheric grid. This will need to be set by
   other oceans if ice dynamics is to be used.

 - MODEL_COM I have added an io_rsf iaction variable: irsficnt which
   will be used for starting a model with the same rsf format, but from a
   rsf file without tracers. (NOT YET FULLY IMPLEMENTED)